util/client_requirements.txt
requests == 2.31.0
pynacl == 1.5.0